2810083 EntityProxyInterceptor refresh issue
I added a staleTimeoutMap to the domain object wrapper to map an
entity class to a timeout used to determine in an object should be
reloaded. I set the timeout for dictionary entities to 5 minutes
and all other entities to 1 second.

I added lastRefreshTime as a constructor param to the entity proxy
interceptor so that the time an object is loaded can be supplied when
the proxy is created. The domain object wrapper sets the time to the
current time for new objects and the time a parent object was loaded 
when wrapping entities accessed from an already wrapped entity.