https://sourceforge.net/p/requel/bugs/16/
A bunch of jaxb compile issues. I added jaxb-impl library and fix imports. I also switched from using SAXException2 to SAXException as the former doesn't exist in the newer version.

Moved more .property files into the resource folder

I made a couple of changes to the Echo2 code to fix a mime-type issue and an old drag and drop issue with echopointng in extras web container both changes were "pull-requested" on github to the echo2 sources. all versions ending in "a" are local and loaded in the script install-lib-mvn.sh