https://sourceforge.net/p/requel/bugs/16/
I deleted a bunch of the files in conf that were part of the ant build configuration.

In JpaDictionaryRepository I switched from using URI based resource access to stream/reader based, as the file is now being access from inside the jar vs. on the file system and the URI constructor was throwing an IllegalArgumentException: URI is not hierarchical.

I @Deprecated DatabaseCreationListener and use createDatabaseIfNotExist=true on the jdbc url. the db properties are no longer in use.