https://sourceforge.net/p/requel/bugs/16/
I fixed the integration with Stanford parser and NER, now using version 3.8. I had to rebuild the GrammaticalRelationType enum as Stanford switched to using the Universal Dependencies http://universaldependencies.org/ for grammatical annotation. The code doesn't reference that many so it wasn't that hard to convert. I added a utility class GrammaticalRelationTypeBuilder that generated the enum entries based on the Stanford entries in the class UniversalEnglishGrammaticalRelations.

In my StanfordLexicalizedParser wrapper class I had to change parse() to handle a fake ROOT node referenced by a root dependency, the index() of the dependencies starts with the first word as 1 so I normally subtract 1 to get the list index of the word, but the ROOT has index of 0 so I handle it specially. It doesn't reference a word in the sentence.

I was able to include the nlp models via a maven dependency. I need to remove the old ones.