https://sourceforge.net/p/requel/bugs/16/
To get the file upload UI to appear I had to set the X-Frame-Options to samedomain in the Application inner class WebSecurityConfig by adding .and().headers().frameOptions().sameOrigin() to the HttpSecurity passed into configure(). I could then upload an xslt in the document tab. Project import is failing with a NoSuchClassException()

I change the xml name space from "http://www.people.fas.harvard.edu/~rregan/requel" to "http://www.rreganjr.com/requel" and I fixed the classes to be under com.rreganjr.requel for the entityType attribute in the projectPermission element in Requel.xml. That didn't fix the import