https://github.com/rreganjr/Requel/issues/2
In UserImpl the password encryption algorithm is now a property so that it can be updated periodically and automatically when a user resets a password or logs in to the system and the current algorithm used isn't the preferred algorithm. The preferred algorithm is defined as a static final property on the UserImpl class versus being configurable so that it can't be accidentally configured with an insecure algorithm. The intent is that new version will update the UserImpl class with security improvements and when deployed users' passwords will be re-encrypted as they use the system. The preferred algorithm is set to PBKDF2WITHHMACSHA512, which is a private key password encoder that supports an iteration count. The iteration count is also a property and the preferred value is set to 50,000 iterations. PBKDF2 is the minimum recommended by OWASP, but also the only recommended algorithm built-in to Java.

There is now also a password salt property generated using a SecureRandom SHA1PRNG it is reset when upgrading the encryption or resetting the password.

There is now password validation. A password must be between 1 and 128 characters long and can't be only whitespace.

There is now a UserImplTest that covers all the new password changes plus most of the original UserImpl methods.

Signed-off-by: rreganjr <rreganjr@acm.org>