https://github.com/rreganjr/Requel/issues/2
In UserImpl getHashedPassword() was changed to getEncryptedPasword(), which caused a new column encrypted_password to be added, that when upgrading an older database causes the passwords originally stored in hashed_password to be ignored and users could not login. The column name is now fixed to hashed_password via the @Column annotation on getEncryptedPassword().

the version in pom.xml was incremenbed to 1.0.2 to match the new milestone.

Signed-off-by: rreganjr <rreganjr@acm.org>