https://github.com/rreganjr/Requel/issues/3
I created the file, but it doesn't work. I've made the web app depend on the db, but the db is going through an init process and the web app fails to start with a CommunicationsException: Communications link failure. I don't know if my connection setup is wrong or if it is a timing issue because the database seems to be doing something different than when I do this using docker command line.

Signed-off-by: rreganjr <rreganjr@acm.org>