https://github.com/rreganjr/Requel/issues/3
I was able to get the compose working via fixing the environment settings for the web app part.

The out of memory problems thanks to https://stackoverflow.com/questions/28327620/difference-between-java-options-java-tool-options-and-java-opts/40736821#40736821
Passing database config thanks to https://aws.amazon.com/blogs/devops/deploying-a-spring-boot-application-on-aws-using-aws-elastic-beanstalk/

Using SPRING_APPLICATION_JSON was ugly and I kept getting connection errors, although maybe that was because the database instance in the docker machine was shutdown wrong and recovering slow and not available when the web app starts.

Signed-off-by: rreganjr <rreganjr@acm.org>