https://github.com/rreganjr/Requel/issues/4
I had to add passwordEncryptingIterations to the @XmlType() annotation propOrder property on UserImpl. An NPE was also thrown when JaxB fails if null is supplied to analyze() on the projectAssistant. I changed it to check for project being non-null before calling the analyse methods, otherwise it just returns without doing anything.

Signed-off-by: rreganjr <rreganjr@acm.org>