https://github.com/rreganjr/Requel/issues/7 https://github.com/rreganjr/Requel/issues/8 https://github.com/rreganjr/Requel/issues/13 https://github.com/rreganjr/Requel/issues/15 branch 1.0.3
I added exception module and moved some exceptions to it, I removed RequelException and just use ApplicationException isntead. most are at the root level com.rreganjr.
I added initializer module and moved the SystemInitializer and AbstractSystemInitializer to it. I split this because both nlp and requel use it.
I added repository module and moved the repository package to it. I split this because both nlp and requel use it.
I added nlp module and moved the nlp package to it.

I removed the UserSet/UserSetImpl and replaced with Set<User> and TreeSet<User>

Signed-off-by: rreganjr <rreganjr@acm.org>