https://github.com/rreganjr/Requel/issues/7 branch 1.0.3
I moved the nlp ui package out of requel and into a new nlp-ui module. I enabled the nlp nav panel in uiMainConfig.xml but the ParserPanel won't open. I have no idea how to register the panel to appear for the open event. this ui was only for playing with the parser directly and not for working with requirements.

Signed-off-by: rreganjr <rreganjr@acm.org>