https://github.com/rreganjr/Requel/issues/17 branch 1.0.3
new module annotatable-mapping to hold named @AnyMetaDef name = "annotatables" that references the requel-project module so that project entites can be MetaValue mapped to the annotatables collection on AbstractAnnotation without a circular dependency between annotation and project modules. Not sure this is working as I had a lot of cyclic dependencies via the iml files, although some of the references seemed suspect and I did some manual cleanup.

Signed-off-by: rreganjr <rreganjr@acm.org>