https://github.com/rreganjr/Requel/issues/17 branch 1.0.3
hopefully I got the dependencies squared away. annotation module doesn't depend on annotatable-mapping in pom.xml as it caused a cycle with annotatable-mapping -> requel-project -> annotation. I only need the AnyMetaDef at runtime so I added it as a runtime dependency to the main requel module.

it sucks that I needed to add .idea to project as it has lots of crap, but sharing building etc. between different dev environments required this or maybe hours of reading to figure out how to properly set it up.