https://github.com/rreganjr/Requel/issues/32
migrating from javax to jakarta
jakarta.annotation, JAXB, persistence

note leaving the javax libraries in pom so that the spring boot that depends on them still works even though this code is using the newer package names.

at this point the next step is to move to boot 3 + hibernate 6 and finish port to the new packages.