https://github.com/rreganjr/Requel/issues/32
app compiles and runs.

phase 3 step 2 align configuration and properties.

mostly updating xml schema location in the config xml to use https and one change in the properties to use driver-class-name instead of driverClassName.

in JpaDictionaryRepository removed some unused imports, commented out imports for the old spell checker and in getConceptCount the query returns a Long instead of BigInteger, this was throwing an exception.