https://github.com/rreganjr/Requel/issues/32
app compiles and runs.

fix issue related to duplicate keys, AbstractAnnotation, ArgumentImpl and PositionImpl were using 	@GeneratedValue(strategy = GenerationType.AUTO) and needed to be switched to @GeneratedValue(strategy = GenerationType.IDENTITY)

Another issue with a change that is probably over board (thanks AI!) is in JpaDictionaryRepository, switching to using hibernate NativeQUery instead of the JPA Query interface, but I don't think it is needed, the real issue is sense.getId() returns a SenseId object not a long, old hibernate could pick out the part it needed but not hibernate 6, the unwrapNativeQuery and using bindLong probably was uneeded just changing setParameter to use ssense.getId().getSynsetid() was probably enough.