https://github.com/rreganjr/Requel/issues/9
ProjectXmlRoundTripIT is now working, uses an in memory database and "test" profile to isolate from real data.

to get around the XML missing the user data because of a proxy, UserStakeholderImpl getUser is annotated with @XmlJavaTypeAdapter(User2UserImplAdapter.class) and making the all the private setters/getters in UserImpl public so the serializer has access to them.

The other fix relates to the annotations, the project.xsd was updated so that at the project level there are two sets of annotation elements, one contains all the annotation elements like notes, issues etc defined through out the project, then there is an annotations section that has annotationRefs for annotations on the project itself, I fell like this is a technical issue, and not really a user requirement to have annotations on a project. I think the real fix is the afterUnmarshal and afterUnmarshalFixupChildren