https://github.com/rreganjr/Requel/issues/9
there is still an issue with export/import I added a top level scenario not associated with a use case through the user interface and the scenario is in the xml file, but when I import it doesn't appear in the UI, if I export that project, the scenario is not in the xml file, meaning the import is losing it.

I've updated the sample with a top level scenario for testing.