https://github.com/rreganjr/Requel/issues/9
updated the configuration of jaxb to remove dependencies on annotation classes. The JaxbAdapterConfigurer adds the DateAdapter and User2UserImplAdapter and the com.rreganjr.requel.annotation.impl.package-info, which is in the requel-app module, registers them for annotation classes. The Import/ExportProjectCommandImpl configure the marshaller/unmarshaller and use The UnmarshallerListenerFactory to build the UnmarshallerListener.

now the annotation package can be split out to its own module without depending on anything in requel-app.