https://github.com/rreganjr/Requel/issues/9
I removed this from a bunch of classes as it was only used for javadoc

import com.rreganjr.requel.utils.jaxb.UnmarshallerListener;

in the javadoc I used the full package name

@see com.rreganjr.requel.utils.jaxb.UnmarshallerListener

I moved JAXBAnnotatablePatcher to the annotation package.

I moved JAXBCreatedEntityPatcher to the user-jpa module because it depends on the UserRepository

Now only the AbstractAnnotation depends on JAXBAnnotationGroupedByPatcher, which is outside the package and still in requel-app. that's a more complicated move because it has dependencies on the requel domain objects ProjectOrDomain and ProjectOrDomainEntity