https://github.com/rreganjr/Requel/issues/9
fixed issue with positions where actions like add a term or actor didn't work, it ended up being the PositionAssembler for the import was just creating  PositionImpl, added ProjectPositionAssembler that builds the proper type from the type in the xml. First tried to fix it via the AnnotationCommandFactoryImpl newResolveIssueCommand() not sure if all those changes are needed. In PositionImpl changed getTYpe() to public so the AnnotationCommandFactoryImpl can build the right type. In PositionImportDraft need to set the positionType from the builder so the ProjectPositionAssembler can figure out the right type on import.

there was an issue in the DeleteActorCommandImpl where it failed because annotations still referenced the actor, seems like something that should be fixable with a jpa cascade, but a native sql delete and some flushes fixed it.

not sure how, but OrganizationImpl wasn't generating ids correctly, updated to IDENTITY type.

Also to fix import errors related to Organizations not being created correctly, updates to resolveOrganization to cache and return an existing organization, also make sure the user has a role, some of the xml is formatted without a role.