I hit more db issues with ids, I removed the migrations, made the init one clean with AUTO_INCREMENT only and no sequence tables, then added a second migration for any old/existing databases to remove sequence tables, update keys to be AUTO_INCREMENT and change the annotation discriminators in one file. I want to start clean.
updated the readme with notes on how to start requel the first time with the new version and an old database so the right flyway script is run.

moved DEPLOY.md to RELEASE_GUIDE.md as that is what it is, plus updated with better notes to follow, so I don't do a bad release again.